Add default for run result env keys column values
As part of a CI check I run `dbt compile` but I get an error:
```
In get_column_values(): relation staging.dbt_artifacts.stg_dbt__run_results_env_keys does not exist and no default value was provided.
```
Although the docs for `get_column_values` state that the default is `[]` the default in the code is actually None. As such the code can't be compiled if the table doesn't exist.